'feat: refactor footer and add cart view'
