+++ client/index.js
@@ -1,9 +1,9 @@
 import './index.scss'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
+import { Provider } from 'react-redux'
 import store from './store'
-import Routes from './routes'
+import Routes from './routes.jsx'
 
 // establishes socket connection
 import './socket'
